+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 - crash when a member gets added to a group and "View Group" dialog is open #5111
 - show appropriate state in AddMemberDialog #5114
+- tauri: fix: ignore `dcnotification:` deep-link when the app is already running
 
 <a id="1_58_2"></a>